requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
pandas
faiss-cpu
hdbscan
tqdm